// frontend/src/app/types/api.d.ts
export type components = {
    schemas: {
// ...
        EmployeeDataResponse: components["schemas"]["BaseResponse"] & {
            data?: components["schemas"]["Employee"];
        };
        EmployeeGetInput: {
            employeeIds: number[];
            fields?: ("id" | "identityNumber" | "phone" | "salary" | "name" | "email" | "role" | "active" | "creationDate")[];
        };
        EmployeeGetResponse: components["schemas"]["BaseResponse"] & {
            data?: components["schemas"]["Employee"][];
        };
// ...
        EmployeeListResponse: components["schemas"]["BaseResponse"] & {
            data?: components["schemas"]["Employee"][];
            page?: number;
            total?: number;
        };
// ...
        ResultResponse: components["schemas"]["BaseResponse"] & {
            result?: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: {
// ...
        EmployeeCreateRequest: {
            content: {
                "application/json": components["schemas"]["EmployeeCreateInput"];
            };
        };
        EmployeeDeleteRequest: {
            content: {
                "application/json": {
                    employeeId: number;
                };
            };
        };
        EmployeeGetRequest: {
            content: {
                "application/json": components["schemas"]["EmployeeGetInput"];
            };
        };
        EmployeeListRequest: {
            content: {
                "application/json": components["schemas"]["EmployeeListInput"];
            };
        };
        EmployeeRestoreRequest: {
            content: {
                "application/json": {
                    employeeId: number;
                };
            };
        };
        EmployeeUpdateRequest: {
            content: {
                "application/json": components["schemas"]["EmployeeUpdateInput"];
            };
        };
// ...
    };
    headers: never;
    pathItems: never;
};